Added Window.cursor_rect. FileFinder made. RPG classes made in c++ with ruby API. Fixed memory leaks in graphical objects. Fixed problem with msgbox and fps. Added marshal _load and _dump for objects. Input buttons are now detected. Other minor fixes.
git-svn-id: https://argss.svn.sourceforge.net/svnroot/argss/trunk@25 a70b67f4-0116-4799-9eb2-a6e6dfe7dbda

+++ src/input.h
@@ -38,10 +38,10 @@ namespace Input {
 	void Init();
 	void Update();
 	void ClearKeys();
-	bool IsPressed(int button);
-	bool IsTriggered(int button);
-	bool IsRepeated(int button);
-	bool IsReleased(int button);
+	bool IsPressed(unsigned long button);
+	bool IsTriggered(unsigned long button);
+	bool IsRepeated(unsigned long button);
+	bool IsReleased(unsigned long button);
 	unsigned long GetPressed();
 	unsigned long GetTriggered();
 	unsigned long GetRepeated();
